requirements.txt
alembic==0.8.5
amqp==1.4.9
anyjson==0.3.3
appdirs==1.4.0
appnope==0.1.0
billiard==3.3.0.22
bitarray==0.8.1
bleach==2.1.4
celery==3.1.20
click==6.6
colorama==0.3.7
cycler==0.9.0
dataset==0.6.2
datasketch>=1.6.0
decorator==4.0.11
defusedxml==0.5.0
Django==1.10.5
editdistance==0.2
elasticsearch==7.0.0
elasticsearch-dsl==7.2.1
entrypoints==0.2.3
Flask==1.0.2
Flask-Cors==2.1.2
gnureadline==6.3.3
html5lib==0.999999999
ipykernel==4.2.1
ipython==4.0.1
ipython-genutils==0.1.0
ipywidgets==4.1.1
isodate==0.5.4
itsdangerous==0.24
java-random==1.0
Jinja2==2.10
jsonschema==2.5.1
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==4.0.3
jupyter-core==4.4.0
kombu==3.0.33
line-profiler==1.0
Mako==1.0.4
MarkupSafe==0.23
matplotlib==1.5.1
mistune==0.8.3
mock==2.0.0
nbconvert==5.4.0
nbformat==4.4.0
networkx==1.11
nltk==3.1
normality==0.2.4
notebook==5.7.0
numpy==1.13.3
ontospy==1.7.3
packaging==16.8
pandas==0.19.2
pandocfilters==1.4.2
path.py==8.1.2
pbr==1.10.0
pdb==0.1
pexpect==4.0.1
pickleshare==0.5
prometheus-client==0.3.1
psycopg2==2.7.1
ptyprocess==0.5
pyfiglet==0.7.5
Pygments==2.0.2
pymongo==3.2
pyparsing==2.0.6
python-dateutil==2.4.2
python-editor==1.0
python-gnupg==0.3.9
pytz==2015.7
PyYAML>=5.1
pyzmq==17.1.2
qtconsole==4.1.1
rdflib==4.2.2
rdflib-jsonld==0.4.0
readline==6.2.4.1
redis==2.10.5
requests==2.13.0
scikit-learn==0.19.1
scipy==1.0.0
Send2Trash==1.5.0
simplegeneric==0.8.1
six==1.10.0
snakeviz==0.4.2
SPARQLWrapper==1.8.0
SQLAlchemy==1.0.12
terminado==0.8.1
testpath==0.3.1
tornado==4.3
tqdm==4.26.0
traitlets==4.3.2
urllib3==1.21.1
uWSGI==2.0.12
webencodings==0.5
Werkzeug==0.14.1
yappi==0.94
fire==0.1.3
dataclasses==0.6
neo4j==1.7.1

# bq_profiler dependencies
google-cloud-bigquery>=3.0.0
google-auth>=2.0.0
